step-by-step tests bug fix and refactor
